[OPIK-7246] separate concerns: buffer-write service, MySQL-only ProjectService, drain in job
Refactor per review:
- ProjectServiceImpl now only does the MySQL project update (recordLastUpdatedTrace);
  dropped the Redis/config dependencies, the buffering branch and the drain. Reverted to
  the Lombok constructor (no @Config here anymore).
- New ProjectLastUpdatedTraceBufferService (domain) owns the ingestion-path Redis buffer
  write (addIfGreater) and the enabled/disabled decision; ProjectEventListener uses it.
- ProjectLastUpdatedFlushJob owns ProjectLastUpdatedFlushConfig and the Redis read/drain
  (renamenx snapshot + paged entryRange/removeAll + parse) and calls ProjectService for the
  MySQL write.
- Tests: add ProjectLastUpdatedTraceBufferServiceTest; move drain coverage into
  ProjectLastUpdatedFlushJobTest; remove ProjectServiceLastUpdatedTraceTest.
